refactor(menu): split door_handler.go by execution backend
door_handler.go (//go:build !windows) was 1,206 lines. Extract the two
large execution backends into sibling files carrying the same build tag
(pure code movement, no behavior change):

- door_handler_dos.go    - DOS/dosemu door path (dropfile name, batch file,
                            dosemurc, executeDOSDoor)
- door_handler_native.go - executeNativeDoor

door_handler.go drops to ~460 lines (door lifecycle/dispatch + the
RUN: handlers runListDoors/runOpenDoor/runDoorInfo). Verified on both
GOOS=darwin/linux (!windows) and GOOS=windows. go vet clean, 274 menu
tests pass (race-clean), 1428 total across 49 packages.
