+++ internal/menu/door_handler.go
@@ -3,7 +3,6 @@
 package menu
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -14,764 +13,19 @@ import (
 	"path/filepath"
 	"sort"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/ansi"
 	"github.com/ViSiON-3/vision-3-bbs/internal/config"
 	"github.com/ViSiON-3/vision-3-bbs/internal/terminalio"
 	"github.com/ViSiON-3/vision-3-bbs/internal/user"
-	"github.com/creack/pty"
-	"golang.org/x/term"
 )
 
 // --- Batch File Generator ---
 
-// dosDropfileName returns the DOS filename for the configured dropfile type.
-func dosDropfileName(dropfileType string) string {
-	switch strings.ToUpper(strings.TrimSpace(dropfileType)) {
-	case "DOOR32.SYS":
-		return "DOOR32.SYS"
-	case "CHAIN.TXT":
-		return "CHAIN.TXT"
-	case "DORINFO1.DEF":
-		return "DORINFO1.DEF"
-	default:
-		return "DOOR.SYS"
-	}
-}
-
-// writeBatchFile generates EXTERNAL.BAT for dosemu2 execution.
-func writeBatchFile(ctx *DoorCtx, batchPath string) error {
-	log.Printf("INFO: Writing batch file: %s", batchPath)
-	crlf := "\r\n"
-
-	var b strings.Builder
-	b.WriteString("@echo off" + crlf)
-	// Ensure dosemu commands (exitemu, etc.) are on the PATH.
-	// D:\dosemu is the standard location for dosemu2-cmds.
-	b.WriteString("SET PATH=%PATH%;D:\\dosemu" + crlf)
-	b.WriteString("c:" + crlf)
-	// Load FOSSIL driver if configured (needed by most BBS door games for serial I/O)
-	if ctx.Config.FossilDriver != "" {
-		b.WriteString(ctx.Config.FossilDriver + crlf)
-	}
-
-	// Clear screen — the PTY bridge detects this ESC[2J sequence to know
-	// boot is done, discarding all prior DOS boot text.
-	b.WriteString("cls" + crlf)
-
-	// Auto-cd to the door's working directory if configured
-	if ctx.Config.WorkingDirectory != "" {
-		// Switch drive letter first (e.g. "D:"), then cd with quotes for spaces
-		if len(ctx.Config.WorkingDirectory) >= 2 && ctx.Config.WorkingDirectory[1] == ':' {
-			b.WriteString(ctx.Config.WorkingDirectory[:2] + crlf)
-		}
-		b.WriteString(fmt.Sprintf("cd \"%s\"", ctx.Config.WorkingDirectory) + crlf)
-	}
-
-	for _, cmd := range ctx.Config.Commands {
-		processed := strings.ReplaceAll(cmd, "{NODE}", ctx.NodeNumStr)
-		for key, val := range ctx.Subs {
-			processed = strings.ReplaceAll(processed, key, val)
-		}
-		// In DOS, running a .BAT without CALL replaces the current batch
-		// context — execution never returns, so exitemu would never run.
-		// Prepend CALL when the first token (the executable) ends with .BAT,
-		// so control returns to EXTERNAL.BAT after the called batch finishes.
-		trimUpper := strings.ToUpper(strings.TrimSpace(processed))
-		if !strings.HasPrefix(trimUpper, "CALL ") {
-			firstToken := strings.Fields(trimUpper)
-			if len(firstToken) > 0 && strings.HasSuffix(firstToken[0], ".BAT") {
-				processed = "CALL " + processed
-			}
-		}
-		b.WriteString(processed + crlf)
-	}
-
-	b.WriteString("exitemu" + crlf)
-
-	return os.WriteFile(batchPath, []byte(b.String()), 0600)
-}
-
-// writeDosemurc generates a per-node dosemurc config file that points
-// $_hdimage at the correct drive_c path and allows lredir to the node dir.
-func writeDosemurc(ctx *DoorCtx, rcPath, driveCPath, nodePath string) error {
-	log.Printf("INFO: Writing dosemurc: %s (drive_c=%s)", rcPath, driveCPath)
-
-	// Start from the user's custom config if one exists, otherwise use defaults
-	baseConfig := ctx.Config.DosemuConfig
-	var content string
-	if baseConfig != "" {
-		data, err := os.ReadFile(baseConfig)
-		if err != nil {
-			return fmt.Errorf("failed to read base dosemurc %s: %w", baseConfig, err)
-		}
-		content = string(data)
-	} else {
-		// Read the default ~/.dosemu/.dosemurc if it exists
-		homeDir, _ := os.UserHomeDir()
-		defaultRc := filepath.Join(homeDir, ".dosemu", ".dosemurc")
-		if data, err := os.ReadFile(defaultRc); err == nil {
-			content = string(data)
-		}
-	}
-
-	// Strip settings that we manage ourselves to avoid conflicts.
-	// $_hdimage is handled via --Fdrive_c CLI argument instead, because
-	// $_hdimage "+0" always resolves to ~/.dosemu/drive_c.
-	// $_video must not be set so the -t flag can select Video_term.
-	// $_com1 is managed below depending on whether a FOSSIL driver is used.
-	var filtered strings.Builder
-	for _, line := range strings.Split(content, "\n") {
-		trimmed := strings.TrimSpace(line)
-		if strings.HasPrefix(trimmed, "$_com1") ||
-			strings.HasPrefix(trimmed, "$_hdimage") ||
-			strings.HasPrefix(trimmed, "$_video") ||
-			strings.HasPrefix(trimmed, "$_vga") {
-			continue
-		}
-		filtered.WriteString(line)
-		filtered.WriteString("\n")
-	}
-	content = filtered.String()
-
-	// Append the critical settings
-	content += fmt.Sprintf("\n# Generated by ViSiON/3 for node %d\n", ctx.NodeNumber)
-	content += fmt.Sprintf("$_lredir_paths = \"%s\"\n", nodePath)
-
-	if ctx.Config.FossilDriver != "" {
-		// FOSSIL mode: doors communicate via serial port through a FOSSIL driver.
-		// $_com1 = "virtual" maps COM1 to the PTY for serial I/O.
-		// $_video = "none" suppresses all screen output (no boot text on PTY).
-		// The -t flag is NOT used in FOSSIL mode — only serial data reaches the PTY.
-		content += "$_com1 = \"virtual\"\n"
-		content += "$_video = \"none\"\n"
-	} else {
-		// Terminal mode: doors write directly to the DOS screen (INT 10h).
-		// -t translates screen writes to ANSI on the PTY. No serial port needed.
-		content += "$_term_color = (on)\n"
-		content += "$_term_esc_char = (27)\n"
-	}
-
-	return os.WriteFile(rcPath, []byte(content), 0600)
-}
-
 // --- DOS Door Executor ---
 
-// executeDOSDoor launches a DOS door program via dosemu2.
-// Uses PTY bridging: dosemu's terminal translator converts INT 10h screen
-// output to ANSI on stdout, which flows through the PTY to the SSH session.
-// Boot text is suppressed by gating PTY output until the cls ESC[2J sequence.
-func executeDOSDoor(ctx *DoorCtx) error {
-	// Determine drive_c path (supports relative paths resolved against BBS root)
-	driveCPath := ctx.Config.DriveCPath
-	if driveCPath == "" {
-		homeDir, err := os.UserHomeDir()
-		if err != nil {
-			return fmt.Errorf("failed to get home directory: %w", err)
-		}
-		driveCPath = filepath.Join(homeDir, ".dosemu", "drive_c")
-	} else if !filepath.IsAbs(driveCPath) {
-		// Resolve relative paths against BBS root (parent of configs dir)
-		bbsRoot := filepath.Dir(ctx.Executor.RootConfigPath)
-		driveCPath = filepath.Join(bbsRoot, driveCPath)
-	}
-
-	// Determine dosemu binary path — use dosemu2.bin directly to bypass
-	// the bash wrapper script which mangles backslashes via eval.
-	dosemuPath := ctx.Executor.ServerCfg.DosemuPath
-	if dosemuPath == "" {
-		dosemuPath = "/usr/libexec/dosemu2/dosemu2.bin"
-	}
-
-	// Create per-node temp directory inside drive_c
-	nodeDir := fmt.Sprintf("temp%d", ctx.NodeNumber)
-	nodePath := filepath.Join(driveCPath, "nodes", nodeDir)
-	if err := os.MkdirAll(nodePath, 0700); err != nil {
-		return fmt.Errorf("failed to create node directory %s: %w", nodePath, err)
-	}
-
-	// Generate all dropfiles
-	if err := generateAllDropfiles(ctx, nodePath); err != nil {
-		return fmt.Errorf("failed to generate dropfiles: %w", err)
-	}
-	defer cleanupDropfiles(nodePath)
-
-	// Populate placeholders for batch file substitution
-	dosNodeDir := fmt.Sprintf("C:\\NODES\\%s", strings.ToUpper(nodeDir))
-	dropfileName := dosDropfileName(ctx.Config.DropfileType)
-	ctx.Subs["{NODEDIR}"] = nodePath
-	ctx.Subs["{DROPFILE}"] = filepath.Join(nodePath, dropfileName)
-	ctx.Subs["{DOSNODEDIR}"] = dosNodeDir
-	ctx.Subs["{DOSDROPFILE}"] = dosNodeDir + "\\" + dropfileName
-
-	// Write batch file
-	batchPath := filepath.Join(nodePath, "EXTERNAL.BAT")
-	if err := writeBatchFile(ctx, batchPath); err != nil {
-		return fmt.Errorf("failed to write batch file: %w", err)
-	}
-
-	// Generate per-node dosemurc with correct drive_c path
-	nodeRcPath := filepath.Join(nodePath, "dosemurc")
-	if err := writeDosemurc(ctx, nodeRcPath, driveCPath, nodePath); err != nil {
-		return fmt.Errorf("failed to write dosemurc: %w", err)
-	}
-
-	// Build dosemu command
-	dosBatchPath := fmt.Sprintf("C:\\NODES\\%s\\EXTERNAL.BAT", strings.ToUpper(nodeDir))
-	logPath := filepath.Join(nodePath, "dosemu_boot.log")
-
-	// Always use the generated per-node config (writeDosemurc already merges
-	// ctx.Config.DosemuConfig as a base template when set)
-	rcPath := nodeRcPath
-
-	// Build dosemu args: FOSSIL mode uses serial I/O (no -t needed),
-	// terminal mode uses -t for screen-to-ANSI translation.
-	args := []string{}
-	if ctx.Config.FossilDriver == "" {
-		args = append(args, "-t")
-	}
-	args = append(args,
-		"-f", rcPath,
-		"--Fdrive_c", driveCPath,
-		"-E", dosBatchPath,
-		"-o", logPath,
-	)
-
-	log.Printf("INFO: Node %d: Launching DOS door '%s' via dosemu2: %s %v", ctx.NodeNumber, ctx.DoorName, dosemuPath, args)
-
-	cmd := exec.Command(dosemuPath, args...)
-	cmd.Dir = driveCPath
-	cmd.Env = os.Environ()
-	cmd.Env = append(cmd.Env, "DOSEMU_QUIET=1")
-
-	// Start dosemu with a PTY for terminal-mode I/O. Dosemu translates DOS
-	// screen output (INT 10h) to ANSI via $_term_color and writes to stdout.
-	// The PTY bridges this to the SSH session.
-	ptmx, err := pty.StartWithSize(cmd, &pty.Winsize{Rows: 25, Cols: 80})
-	if err != nil {
-		return fmt.Errorf("failed to start dosemu2 with pty: %w", err)
-	}
-
-	// Set PTY master to raw mode for clean passthrough of CP437 bytes
-	fd := int(ptmx.Fd())
-	if oldState, err := term.MakeRaw(fd); err == nil {
-		defer term.Restore(fd, oldState)
-	}
-
-	// Set up a read interrupt so we can cleanly stop the input goroutine
-	// when the door exits, preventing it from consuming the next keypress.
-	// Sessions that support SetReadInterrupt (SSH) get clean cancellation;
-	// others (telnet) fall back to the old behavior where the input goroutine
-	// exits on its own when ptmx is closed and the next write fails.
-	readInterrupt := make(chan struct{})
-	hasInterrupt := false
-	if ri, ok := ctx.Session.(interface{ SetReadInterrupt(<-chan struct{}) }); ok {
-		ri.SetReadInterrupt(readInterrupt)
-		defer ri.SetReadInterrupt(nil)
-		hasInterrupt = true
-	}
-
-	// Bidirectional I/O: SSH session ↔ PTY master
-	//
-	// Two modes depending on whether a FOSSIL driver is configured:
-	//
-	// FOSSIL mode: $_com1="virtual" maps COM1 to the PTY. The door's output
-	// flows through the FOSSIL driver → COM1 → PTY. $_video="none" suppresses
-	// DOS screen output, but dosemu still writes plain-ASCII boot text directly
-	// to the PTY. Gate on the first ESC (0x1B) byte: boot text is plain ASCII,
-	// while door output (via FOSSIL) contains ANSI escape sequences.
-	//
-	// Terminal mode: -t translates DOS screen writes (INT 10h) to ANSI on
-	// the PTY. The output goroutine buffers until it detects ESC[2J (clear
-	// screen from `cls` in EXTERNAL.BAT), discarding DOS boot text before it.
-	useFossil := ctx.Config.FossilDriver != ""
-
-	inputDone := make(chan struct{})
-	outputDone := make(chan struct{})
-	go func() {
-		defer close(inputDone)
-		_, err := io.Copy(ptmx, ctx.Session)
-		if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
-			log.Printf("WARN: Node %d: Error copying session stdin to dosemu PTY: %v", ctx.NodeNumber, err)
-		}
-	}()
-	go func() {
-		defer close(outputDone)
-		buf := make([]byte, 4096)
-
-		if useFossil {
-			// FOSSIL mode: dosemu writes plain-ASCII boot text to the PTY before
-			// the door starts. Door output (via FOSSIL → COM1 → PTY) contains
-			// ANSI escape codes. Gate on the first ESC (0x1B) byte to suppress
-			// boot text while forwarding all door output.
-			log.Printf("DEBUG: Node %d: FOSSIL mode — gating boot text until first ESC byte", ctx.NodeNumber)
-			var accum []byte
-			gated := true
-			for {
-				n, err := ptmx.Read(buf)
-				if n > 0 {
-					if gated {
-						accum = append(accum, buf[:n]...)
-						if idx := bytes.IndexByte(accum, 0x1b); idx >= 0 {
-							gated = false
-							ctx.Session.Write(accum[idx:])
-							accum = nil
-							log.Printf("DEBUG: Node %d: FOSSIL boot text gate opened (ESC byte at offset %d)", ctx.NodeNumber, idx)
-						}
-						if gated && len(accum) > 32768 {
-							log.Printf("WARN: Node %d: FOSSIL gate: no ESC after 32KB, flushing", ctx.NodeNumber)
-							gated = false
-							ctx.Session.Write(accum)
-							accum = nil
-						}
-					} else {
-						ctx.Session.Write(buf[:n])
-					}
-				}
-				if err != nil {
-					if gated && len(accum) > 0 {
-						log.Printf("WARN: Node %d: FOSSIL boot text gate never opened. Accumulated %d bytes", ctx.NodeNumber, len(accum))
-					}
-					if err != io.EOF && !errors.Is(err, os.ErrClosed) {
-						log.Printf("WARN: Node %d: Error copying dosemu PTY to session: %v", ctx.NodeNumber, err)
-					}
-					return
-				}
-			}
-		}
-
-		// Terminal mode: gate until ESC[2J (cls) to suppress boot text
-		clearScreen := []byte("\x1b[2J")
-		var accum []byte
-		gated := true
-
-		for {
-			n, err := ptmx.Read(buf)
-			if n > 0 {
-				if gated {
-					accum = append(accum, buf[:n]...)
-					if idx := bytes.Index(accum, clearScreen); idx >= 0 {
-						// Found clear screen — forward it and everything after
-						gated = false
-						ctx.Session.Write(accum[idx:])
-						accum = nil
-						log.Printf("DEBUG: Node %d: DOS boot text gate opened (clear screen detected)", ctx.NodeNumber)
-					}
-					// Safety: if we've accumulated too much without finding cls, flush it
-					if gated && len(accum) > 32768 {
-						log.Printf("WARN: Node %d: Clear screen not found after 32KB, flushing", ctx.NodeNumber)
-						gated = false
-						ctx.Session.Write(accum)
-						accum = nil
-					}
-				} else {
-					ctx.Session.Write(buf[:n])
-				}
-			}
-			if err != nil {
-				if gated && len(accum) > 0 {
-					log.Printf("WARN: Node %d: DOS boot text gate never opened. Accumulated %d bytes. First 200: %q", ctx.NodeNumber, len(accum), accum[:min(200, len(accum))])
-				}
-				if err != io.EOF && !errors.Is(err, os.ErrClosed) {
-					log.Printf("WARN: Node %d: Error copying dosemu PTY to session: %v", ctx.NodeNumber, err)
-				}
-				return
-			}
-		}
-	}()
-
-	// Wait for dosemu to exit, then cleanly shut down I/O goroutines
-	cmdErr := cmd.Wait()
-	log.Printf("DEBUG: Node %d: dosemu2 process exited for door '%s'", ctx.NodeNumber, ctx.DoorName)
-
-	// Interrupt the input goroutine's blocked Read() so it exits without
-	// consuming the user's next keypress, then close the PTY.
-	close(readInterrupt)
-	if hasInterrupt {
-		<-inputDone
-	}
-	ptmx.Close()
-	<-outputDone
-
-	if cmdErr != nil {
-		log.Printf("ERROR: Node %d: DOS door '%s' failed: %v", ctx.NodeNumber, ctx.DoorName, cmdErr)
-	} else {
-		log.Printf("INFO: Node %d: DOS door '%s' completed successfully", ctx.NodeNumber, ctx.DoorName)
-	}
-
-	// Run cleanup while dropfiles/node dirs still exist (before deferred cleanup fires)
-	executeCleanup(ctx)
-
-	return cmdErr
-}
-
 // --- Native Door Executor ---
-
-// executeNativeDoor runs a native (non-DOS) door program.
-// This is extracted from the original inline DOOR: handler in executor.go.
-func executeNativeDoor(ctx *DoorCtx) error {
-	doorConfig := ctx.Config
-
-	// --- Dropfile Generation ---
-	// Must happen before arg substitution so {DROPFILE} and {NODEDIR} are available.
-	var dropfilePath string
-	dropfileDir := "."
-	if doorConfig.WorkingDirectory != "" {
-		dropfileDir = doorConfig.WorkingDirectory
-	}
-
-	// Configurable dropfile location: "node" uses a unique per-node temp directory.
-	// Uses os.MkdirTemp for unique names and defers os.RemoveAll unconditionally
-	// so the directory is always cleaned up, even if no recognized dropfile is generated.
-	dropfileLoc := strings.ToLower(doorConfig.DropfileLocation)
-	if dropfileLoc == "node" {
-		nodeDir, err := os.MkdirTemp("", fmt.Sprintf("vision3_node%d_", ctx.NodeNumber))
-		if err != nil {
-			return fmt.Errorf("failed to create node dropfile directory: %w", err)
-		}
-		defer func() {
-			if err := os.RemoveAll(nodeDir); err != nil {
-				log.Printf("WARN: Failed to remove node dropfile dir %s: %v", nodeDir, err)
-			}
-		}()
-		dropfileDir = nodeDir
-	}
-
-	dropfileTypeUpper := strings.ToUpper(doorConfig.DropfileType)
-
-	if dropfileTypeUpper == "DOOR.SYS" || dropfileTypeUpper == "CHAIN.TXT" || dropfileTypeUpper == "DOOR32.SYS" || dropfileTypeUpper == "DORINFO1.DEF" {
-		dropfilePath = filepath.Join(dropfileDir, dropfileTypeUpper)
-		log.Printf("INFO: Generating %s dropfile at: %s", dropfileTypeUpper, dropfilePath)
-
-		// Use full-format dropfile generators (standard BBS formats)
-		var genErr error
-		switch dropfileTypeUpper {
-		case "DOOR.SYS":
-			genErr = generateDoorSys(ctx, dropfileDir)
-		case "DOOR32.SYS":
-			genErr = generateDoor32Sys(ctx, dropfileDir)
-		case "CHAIN.TXT":
-			genErr = generateChainTxt(ctx, dropfileDir)
-		case "DORINFO1.DEF":
-			genErr = generateDorInfo(ctx, dropfileDir)
-		}
-
-		if genErr != nil {
-			log.Printf("ERROR: Failed to write dropfile %s: %v", dropfilePath, genErr)
-			errMsg := fmt.Sprintf(ctx.Executor.LoadedStrings.DoorDropfileError, ctx.DoorName)
-			if wErr := terminalio.WriteProcessedBytes(ctx.Session.Stderr(), ansi.ReplacePipeCodes([]byte(errMsg)), ctx.OutputMode); wErr != nil {
-				log.Printf("ERROR: Failed writing dropfile creation error message: %v", wErr)
-			}
-			return genErr
-		}
-
-		defer func() {
-			log.Printf("DEBUG: Cleaning up dropfile: %s", dropfilePath)
-			if err := os.Remove(dropfilePath); err != nil {
-				log.Printf("WARN: Failed to remove dropfile %s: %v", dropfilePath, err)
-			}
-		}()
-	}
-
-	// Populate {DROPFILE} and {NODEDIR} now that dropfile generation is done
-	ctx.Subs["{DROPFILE}"] = dropfilePath
-	ctx.Subs["{NODEDIR}"] = dropfileDir
-
-	// Extract command and args from Commands slice (native: [0]=executable, [1:]=args)
-	if len(doorConfig.Commands) == 0 || doorConfig.Commands[0] == "" {
-		return fmt.Errorf("door %q has no command configured", ctx.DoorName)
-	}
-	doorCommand := doorConfig.Commands[0]
-	doorArgs := doorConfig.Commands[1:]
-
-	// Substitute placeholders in command and arguments
-	for key, val := range ctx.Subs {
-		doorCommand = strings.ReplaceAll(doorCommand, key, val)
-	}
-	substitutedArgs := make([]string, len(doorArgs))
-	for i, arg := range doorArgs {
-		newArg := arg
-		for key, val := range ctx.Subs {
-			newArg = strings.ReplaceAll(newArg, key, val)
-		}
-		substitutedArgs[i] = newArg
-	}
-
-	// Substitute in Environment Variables
-	substitutedEnv := make(map[string]string)
-	if doorConfig.EnvironmentVars != nil {
-		for key, val := range doorConfig.EnvironmentVars {
-			newVal := val
-			for subKey, subVal := range ctx.Subs {
-				newVal = strings.ReplaceAll(newVal, subKey, subVal)
-			}
-			substitutedEnv[key] = newVal
-		}
-	}
-
-	// Prepare command — optionally wrap in shell
-	var cmd *exec.Cmd
-	if doorConfig.UseShell {
-		// Use exec "$0" "$@" to preserve argument boundaries and prevent injection
-		shellArgs := append([]string{"-c", `exec "$0" "$@"`, doorCommand}, substitutedArgs...)
-		cmd = exec.Command("/bin/sh", shellArgs...)
-		log.Printf("DEBUG: Node %d: Using shell execution for %q with %d arg(s)", ctx.NodeNumber, doorCommand, len(substitutedArgs))
-	} else {
-		cmd = exec.Command(doorCommand, substitutedArgs...)
-	}
-
-	if doorConfig.WorkingDirectory != "" {
-		cmd.Dir = doorConfig.WorkingDirectory
-		log.Printf("DEBUG: Setting working directory for door '%s' to '%s'", ctx.DoorName, cmd.Dir)
-	}
-
-	// Set environment variables
-	cmd.Env = os.Environ()
-	if len(substitutedEnv) > 0 {
-		for key, val := range substitutedEnv {
-			cmd.Env = append(cmd.Env, fmt.Sprintf("%s=%s", key, val))
-		}
-	}
-
-	// Add standard BBS env vars if not already present
-	envMap := make(map[string]bool)
-	for _, envPair := range cmd.Env {
-		envMap[strings.SplitN(envPair, "=", 2)[0]] = true
-	}
-	if _, exists := envMap["BBS_USERHANDLE"]; !exists {
-		cmd.Env = append(cmd.Env, fmt.Sprintf("BBS_USERHANDLE=%s", ctx.User.Handle))
-	}
-	if _, exists := envMap["BBS_USERID"]; !exists {
-		cmd.Env = append(cmd.Env, fmt.Sprintf("BBS_USERID=%s", ctx.UserIDStr))
-	}
-	if _, exists := envMap["BBS_NODE"]; !exists {
-		cmd.Env = append(cmd.Env, fmt.Sprintf("BBS_NODE=%s", ctx.NodeNumStr))
-	}
-	if _, exists := envMap["BBS_TIMELEFT"]; !exists {
-		cmd.Env = append(cmd.Env, fmt.Sprintf("BBS_TIMELEFT=%s", ctx.TimeLeftStr))
-	}
-
-	// Set LINES and COLUMNS from user's saved preferences (for terminal size detection).
-	// Remove any existing LINES/COLUMNS entries first to ensure our values take precedence.
-	screenHeight := ctx.User.ScreenHeight
-	if screenHeight <= 0 {
-		screenHeight = 25
-	}
-	screenWidth := ctx.User.ScreenWidth
-	if screenWidth <= 0 {
-		screenWidth = 80
-	}
-	filteredEnv := make([]string, 0, len(cmd.Env))
-	for _, e := range cmd.Env {
-		if !strings.HasPrefix(e, "LINES=") && !strings.HasPrefix(e, "COLUMNS=") {
-			filteredEnv = append(filteredEnv, e)
-		}
-	}
-	cmd.Env = append(filteredEnv, fmt.Sprintf("LINES=%d", screenHeight), fmt.Sprintf("COLUMNS=%d", screenWidth))
-	log.Printf("DEBUG: Node %d: Set door env LINES=%d COLUMNS=%d", ctx.NodeNumber, screenHeight, screenWidth)
-
-	// Execute command
-	_, winChOrig, isPty := ctx.Session.Pty()
-	var cmdErr error
-
-	if doorConfig.RequiresRawTerminal && isPty {
-		log.Printf("INFO: Node %d: Starting door '%s' with PTY/Raw mode", ctx.NodeNumber, ctx.DoorName)
-
-		// Set PTY size from user's saved preferences - BEFORE starting the command
-		doorScreenHeight := uint16(25) // default
-		if ctx.User.ScreenHeight > 0 && ctx.User.ScreenHeight <= 65535 {
-			doorScreenHeight = uint16(ctx.User.ScreenHeight)
-		}
-		doorScreenWidth := uint16(80) // default
-		if ctx.User.ScreenWidth > 0 && ctx.User.ScreenWidth <= 65535 {
-			doorScreenWidth = uint16(ctx.User.ScreenWidth)
-		}
-		doorSize := &pty.Winsize{Rows: doorScreenHeight, Cols: doorScreenWidth}
-		log.Printf("DEBUG: Node %d: Starting door with PTY size %dx%d (from user preferences)", ctx.NodeNumber, doorScreenWidth, doorScreenHeight)
-
-		ptmx, err := pty.StartWithSize(cmd, doorSize)
-		if err != nil {
-			cmdErr = fmt.Errorf("failed to start pty for door '%s': %w", ctx.DoorName, err)
-		} else {
-			ctx.Session.Signals(nil)
-			ctx.Session.Break(nil)
-
-			// Drain window resize events but don't apply them - respect user's saved preferences.
-			// resizeStop is closed after cmd.Wait() to prevent this goroutine from leaking.
-			resizeStop := make(chan struct{})
-			go func() {
-				for {
-					select {
-					case win, ok := <-winChOrig:
-						if !ok {
-							return
-						}
-						log.Printf("DEBUG: Node %d: Ignoring SSH resize event %dx%d (keeping user preference %dx%d)",
-							ctx.NodeNumber, win.Width, win.Height, doorScreenWidth, doorScreenHeight)
-					case <-resizeStop:
-						return
-					}
-				}
-			}()
-
-			fd := int(ptmx.Fd())
-			originalState, err := term.MakeRaw(fd)
-			if err != nil {
-				log.Printf("WARN: Node %d: Failed to put PTY into raw mode for door '%s': %v.", ctx.NodeNumber, ctx.DoorName, err)
-			} else {
-				log.Printf("DEBUG: Node %d: PTY set to raw mode for door '%s'.", ctx.NodeNumber, ctx.DoorName)
-			}
-			needsRestore := (err == nil)
-
-			// Set up a read interrupt so we can cleanly stop the input goroutine
-			// when the door exits, preventing it from consuming the next keypress.
-			// Sessions that support SetReadInterrupt (SSH) get clean cancellation;
-			// others (telnet) fall back to the old behavior where the input goroutine
-			// exits on its own when ptmx is closed and the next write fails.
-			readInterrupt := make(chan struct{})
-			hasInterrupt := false
-			if ri, ok := ctx.Session.(interface{ SetReadInterrupt(<-chan struct{}) }); ok {
-				ri.SetReadInterrupt(readInterrupt)
-				defer ri.SetReadInterrupt(nil)
-				hasInterrupt = true
-			}
-
-			inputDone := make(chan struct{})
-			outputDone := make(chan struct{})
-			go func() {
-				defer close(inputDone)
-				_, err := io.Copy(ptmx, ctx.Session)
-				if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
-					// "read interrupted" is expected when we close readInterrupt during shutdown
-					if strings.Contains(err.Error(), "read interrupted") {
-						log.Printf("DEBUG: Node %d: Input goroutine interrupted for door '%s' (expected during shutdown)", ctx.NodeNumber, ctx.DoorName)
-					} else {
-						log.Printf("WARN: Node %d: Error copying session stdin to PTY for door '%s': %v", ctx.NodeNumber, ctx.DoorName, err)
-					}
-				}
-			}()
-			go func() {
-				defer close(outputDone)
-				_, err := io.Copy(ctx.Session, ptmx)
-				if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
-					// "input/output error" on PTY is expected when closing during active read
-					if strings.Contains(err.Error(), "input/output error") {
-						log.Printf("DEBUG: Node %d: Output goroutine I/O error for door '%s' (expected during shutdown)", ctx.NodeNumber, ctx.DoorName)
-					} else {
-						log.Printf("WARN: Node %d: Error copying PTY stdout to session for door '%s': %v", ctx.NodeNumber, ctx.DoorName, err)
-					}
-				}
-			}()
-
-			// Wait for door to exit, then cleanly shut down I/O goroutines
-			cmdErr = cmd.Wait()
-			close(resizeStop)
-			log.Printf("DEBUG: Node %d: Door '%s' process exited", ctx.NodeNumber, ctx.DoorName)
-
-			// Interrupt the input goroutine's blocked Read() so it exits without
-			// consuming the user's next keypress, then restore PTY state and close.
-			close(readInterrupt)
-			if hasInterrupt {
-				<-inputDone
-			}
-
-			// Restore PTY state before closing the file descriptor
-			if needsRestore {
-				log.Printf("DEBUG: Node %d: Restoring PTY mode after door '%s'.", ctx.NodeNumber, ctx.DoorName)
-				if err := term.Restore(fd, originalState); err != nil {
-					log.Printf("ERROR: Node %d: Failed to restore PTY state after door '%s': %v", ctx.NodeNumber, ctx.DoorName, err)
-				}
-			}
-
-			ptmx.Close()
-			<-outputDone
-		}
-	} else if strings.ToUpper(doorConfig.IOMode) == "SOCKET" {
-		// Socket I/O: create a Unix socketpair and pass one end to the door as FD 3.
-		// The other end is bridged bidirectionally to the BBS session.
-		log.Printf("INFO: Node %d: Starting door '%s' with Socket I/O mode", ctx.NodeNumber, ctx.DoorName)
-
-		fds, err := syscall.Socketpair(syscall.AF_UNIX, syscall.SOCK_STREAM, 0)
-		if err != nil {
-			cmdErr = fmt.Errorf("failed to create socketpair for door '%s': %w", ctx.DoorName, err)
-		} else {
-			// fds[0] = BBS side, fds[1] = door side (will become FD 3 in child)
-			bbsSock := os.NewFile(uintptr(fds[0]), "bbs-socket")
-			doorSock := os.NewFile(uintptr(fds[1]), "door-socket")
-
-			cmd.ExtraFiles = []*os.File{doorSock} // child FD 3
-			cmd.Env = append(cmd.Env, "DOOR_SOCKET_FD=3")
-
-			if startErr := cmd.Start(); startErr != nil {
-				bbsSock.Close()
-				doorSock.Close()
-				cmdErr = fmt.Errorf("failed to start door '%s' with socket I/O: %w", ctx.DoorName, startErr)
-			} else {
-				// Parent closes the door's end
-				doorSock.Close()
-
-				// Set up read interrupt for clean shutdown
-				readInterrupt := make(chan struct{})
-				hasInterrupt := false
-				if ri, ok := ctx.Session.(interface{ SetReadInterrupt(<-chan struct{}) }); ok {
-					ri.SetReadInterrupt(readInterrupt)
-					defer ri.SetReadInterrupt(nil)
-					hasInterrupt = true
-				}
-
-				// Bidirectional bridge: session <-> socketpair
-				inputDone := make(chan struct{})
-				outputDone := make(chan struct{})
-				go func() {
-					defer close(inputDone)
-					_, err := io.Copy(bbsSock, ctx.Session)
-					if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
-						if !strings.Contains(err.Error(), "read interrupted") {
-							log.Printf("WARN: Node %d: Socket I/O input error for door '%s': %v", ctx.NodeNumber, ctx.DoorName, err)
-						}
-					}
-				}()
-				go func() {
-					defer close(outputDone)
-					_, err := io.Copy(ctx.Session, bbsSock)
-					if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
-						log.Printf("WARN: Node %d: Socket I/O output error for door '%s': %v", ctx.NodeNumber, ctx.DoorName, err)
-					}
-				}()
-
-				cmdErr = cmd.Wait()
-				log.Printf("DEBUG: Node %d: Door '%s' (socket I/O) process exited", ctx.NodeNumber, ctx.DoorName)
-
-				close(readInterrupt)
-				if hasInterrupt {
-					<-inputDone
-				}
-				bbsSock.Close()
-				<-outputDone
-			}
-		}
-	} else {
-		if doorConfig.RequiresRawTerminal && !isPty {
-			log.Printf("WARN: Node %d: Door '%s' requires raw terminal, but no PTY was allocated.", ctx.NodeNumber, ctx.DoorName)
-		}
-		log.Printf("INFO: Node %d: Starting door '%s' with standard I/O redirection", ctx.NodeNumber, ctx.DoorName)
-
-		cmd.Stdout = ctx.Session
-		cmd.Stderr = ctx.Session
-		cmd.Stdin = ctx.Session
-		cmdErr = cmd.Run()
-
-		// Brief delay to let terminal state settle and prevent double-keypress issues
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	// Run cleanup while dropfiles/node dirs still exist (before deferred cleanup fires)
-	executeCleanup(ctx)
-
-	return cmdErr
-}
 
 // --- Door Cleanup ---
