+++ internal/menu/door_handler_dos.go
@@ -1,0 +1,396 @@
+//go:build !windows
+
+package menu
+
+import (
+	"bytes"
+	"errors"
+	"fmt"
+	"io"
+	"log"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+
+	"github.com/creack/pty"
+	"golang.org/x/term"
+)
+
+// dosDropfileName returns the DOS filename for the configured dropfile type.
+func dosDropfileName(dropfileType string) string {
+	switch strings.ToUpper(strings.TrimSpace(dropfileType)) {
+	case "DOOR32.SYS":
+		return "DOOR32.SYS"
+	case "CHAIN.TXT":
+		return "CHAIN.TXT"
+	case "DORINFO1.DEF":
+		return "DORINFO1.DEF"
+	default:
+		return "DOOR.SYS"
+	}
+}
+
+// writeBatchFile generates EXTERNAL.BAT for dosemu2 execution.
+func writeBatchFile(ctx *DoorCtx, batchPath string) error {
+	log.Printf("INFO: Writing batch file: %s", batchPath)
+	crlf := "\r\n"
+
+	var b strings.Builder
+	b.WriteString("@echo off" + crlf)
+	// Ensure dosemu commands (exitemu, etc.) are on the PATH.
+	// D:\dosemu is the standard location for dosemu2-cmds.
+	b.WriteString("SET PATH=%PATH%;D:\\dosemu" + crlf)
+	b.WriteString("c:" + crlf)
+	// Load FOSSIL driver if configured (needed by most BBS door games for serial I/O)
+	if ctx.Config.FossilDriver != "" {
+		b.WriteString(ctx.Config.FossilDriver + crlf)
+	}
+
+	// Clear screen — the PTY bridge detects this ESC[2J sequence to know
+	// boot is done, discarding all prior DOS boot text.
+	b.WriteString("cls" + crlf)
+
+	// Auto-cd to the door's working directory if configured
+	if ctx.Config.WorkingDirectory != "" {
+		// Switch drive letter first (e.g. "D:"), then cd with quotes for spaces
+		if len(ctx.Config.WorkingDirectory) >= 2 && ctx.Config.WorkingDirectory[1] == ':' {
+			b.WriteString(ctx.Config.WorkingDirectory[:2] + crlf)
+		}
+		b.WriteString(fmt.Sprintf("cd \"%s\"", ctx.Config.WorkingDirectory) + crlf)
+	}
+
+	for _, cmd := range ctx.Config.Commands {
+		processed := strings.ReplaceAll(cmd, "{NODE}", ctx.NodeNumStr)
+		for key, val := range ctx.Subs {
+			processed = strings.ReplaceAll(processed, key, val)
+		}
+		// In DOS, running a .BAT without CALL replaces the current batch
+		// context — execution never returns, so exitemu would never run.
+		// Prepend CALL when the first token (the executable) ends with .BAT,
+		// so control returns to EXTERNAL.BAT after the called batch finishes.
+		trimUpper := strings.ToUpper(strings.TrimSpace(processed))
+		if !strings.HasPrefix(trimUpper, "CALL ") {
+			firstToken := strings.Fields(trimUpper)
+			if len(firstToken) > 0 && strings.HasSuffix(firstToken[0], ".BAT") {
+				processed = "CALL " + processed
+			}
+		}
+		b.WriteString(processed + crlf)
+	}
+
+	b.WriteString("exitemu" + crlf)
+
+	return os.WriteFile(batchPath, []byte(b.String()), 0600)
+}
+
+// writeDosemurc generates a per-node dosemurc config file that points
+// $_hdimage at the correct drive_c path and allows lredir to the node dir.
+func writeDosemurc(ctx *DoorCtx, rcPath, driveCPath, nodePath string) error {
+	log.Printf("INFO: Writing dosemurc: %s (drive_c=%s)", rcPath, driveCPath)
+
+	// Start from the user's custom config if one exists, otherwise use defaults
+	baseConfig := ctx.Config.DosemuConfig
+	var content string
+	if baseConfig != "" {
+		data, err := os.ReadFile(baseConfig)
+		if err != nil {
+			return fmt.Errorf("failed to read base dosemurc %s: %w", baseConfig, err)
+		}
+		content = string(data)
+	} else {
+		// Read the default ~/.dosemu/.dosemurc if it exists
+		homeDir, _ := os.UserHomeDir()
+		defaultRc := filepath.Join(homeDir, ".dosemu", ".dosemurc")
+		if data, err := os.ReadFile(defaultRc); err == nil {
+			content = string(data)
+		}
+	}
+
+	// Strip settings that we manage ourselves to avoid conflicts.
+	// $_hdimage is handled via --Fdrive_c CLI argument instead, because
+	// $_hdimage "+0" always resolves to ~/.dosemu/drive_c.
+	// $_video must not be set so the -t flag can select Video_term.
+	// $_com1 is managed below depending on whether a FOSSIL driver is used.
+	var filtered strings.Builder
+	for _, line := range strings.Split(content, "\n") {
+		trimmed := strings.TrimSpace(line)
+		if strings.HasPrefix(trimmed, "$_com1") ||
+			strings.HasPrefix(trimmed, "$_hdimage") ||
+			strings.HasPrefix(trimmed, "$_video") ||
+			strings.HasPrefix(trimmed, "$_vga") {
+			continue
+		}
+		filtered.WriteString(line)
+		filtered.WriteString("\n")
+	}
+	content = filtered.String()
+
+	// Append the critical settings
+	content += fmt.Sprintf("\n# Generated by ViSiON/3 for node %d\n", ctx.NodeNumber)
+	content += fmt.Sprintf("$_lredir_paths = \"%s\"\n", nodePath)
+
+	if ctx.Config.FossilDriver != "" {
+		// FOSSIL mode: doors communicate via serial port through a FOSSIL driver.
+		// $_com1 = "virtual" maps COM1 to the PTY for serial I/O.
+		// $_video = "none" suppresses all screen output (no boot text on PTY).
+		// The -t flag is NOT used in FOSSIL mode — only serial data reaches the PTY.
+		content += "$_com1 = \"virtual\"\n"
+		content += "$_video = \"none\"\n"
+	} else {
+		// Terminal mode: doors write directly to the DOS screen (INT 10h).
+		// -t translates screen writes to ANSI on the PTY. No serial port needed.
+		content += "$_term_color = (on)\n"
+		content += "$_term_esc_char = (27)\n"
+	}
+
+	return os.WriteFile(rcPath, []byte(content), 0600)
+}
+
+// executeDOSDoor launches a DOS door program via dosemu2.
+// Uses PTY bridging: dosemu's terminal translator converts INT 10h screen
+// output to ANSI on stdout, which flows through the PTY to the SSH session.
+// Boot text is suppressed by gating PTY output until the cls ESC[2J sequence.
+func executeDOSDoor(ctx *DoorCtx) error {
+	// Determine drive_c path (supports relative paths resolved against BBS root)
+	driveCPath := ctx.Config.DriveCPath
+	if driveCPath == "" {
+		homeDir, err := os.UserHomeDir()
+		if err != nil {
+			return fmt.Errorf("failed to get home directory: %w", err)
+		}
+		driveCPath = filepath.Join(homeDir, ".dosemu", "drive_c")
+	} else if !filepath.IsAbs(driveCPath) {
+		// Resolve relative paths against BBS root (parent of configs dir)
+		bbsRoot := filepath.Dir(ctx.Executor.RootConfigPath)
+		driveCPath = filepath.Join(bbsRoot, driveCPath)
+	}
+
+	// Determine dosemu binary path — use dosemu2.bin directly to bypass
+	// the bash wrapper script which mangles backslashes via eval.
+	dosemuPath := ctx.Executor.ServerCfg.DosemuPath
+	if dosemuPath == "" {
+		dosemuPath = "/usr/libexec/dosemu2/dosemu2.bin"
+	}
+
+	// Create per-node temp directory inside drive_c
+	nodeDir := fmt.Sprintf("temp%d", ctx.NodeNumber)
+	nodePath := filepath.Join(driveCPath, "nodes", nodeDir)
+	if err := os.MkdirAll(nodePath, 0700); err != nil {
+		return fmt.Errorf("failed to create node directory %s: %w", nodePath, err)
+	}
+
+	// Generate all dropfiles
+	if err := generateAllDropfiles(ctx, nodePath); err != nil {
+		return fmt.Errorf("failed to generate dropfiles: %w", err)
+	}
+	defer cleanupDropfiles(nodePath)
+
+	// Populate placeholders for batch file substitution
+	dosNodeDir := fmt.Sprintf("C:\\NODES\\%s", strings.ToUpper(nodeDir))
+	dropfileName := dosDropfileName(ctx.Config.DropfileType)
+	ctx.Subs["{NODEDIR}"] = nodePath
+	ctx.Subs["{DROPFILE}"] = filepath.Join(nodePath, dropfileName)
+	ctx.Subs["{DOSNODEDIR}"] = dosNodeDir
+	ctx.Subs["{DOSDROPFILE}"] = dosNodeDir + "\\" + dropfileName
+
+	// Write batch file
+	batchPath := filepath.Join(nodePath, "EXTERNAL.BAT")
+	if err := writeBatchFile(ctx, batchPath); err != nil {
+		return fmt.Errorf("failed to write batch file: %w", err)
+	}
+
+	// Generate per-node dosemurc with correct drive_c path
+	nodeRcPath := filepath.Join(nodePath, "dosemurc")
+	if err := writeDosemurc(ctx, nodeRcPath, driveCPath, nodePath); err != nil {
+		return fmt.Errorf("failed to write dosemurc: %w", err)
+	}
+
+	// Build dosemu command
+	dosBatchPath := fmt.Sprintf("C:\\NODES\\%s\\EXTERNAL.BAT", strings.ToUpper(nodeDir))
+	logPath := filepath.Join(nodePath, "dosemu_boot.log")
+
+	// Always use the generated per-node config (writeDosemurc already merges
+	// ctx.Config.DosemuConfig as a base template when set)
+	rcPath := nodeRcPath
+
+	// Build dosemu args: FOSSIL mode uses serial I/O (no -t needed),
+	// terminal mode uses -t for screen-to-ANSI translation.
+	args := []string{}
+	if ctx.Config.FossilDriver == "" {
+		args = append(args, "-t")
+	}
+	args = append(args,
+		"-f", rcPath,
+		"--Fdrive_c", driveCPath,
+		"-E", dosBatchPath,
+		"-o", logPath,
+	)
+
+	log.Printf("INFO: Node %d: Launching DOS door '%s' via dosemu2: %s %v", ctx.NodeNumber, ctx.DoorName, dosemuPath, args)
+
+	cmd := exec.Command(dosemuPath, args...)
+	cmd.Dir = driveCPath
+	cmd.Env = os.Environ()
+	cmd.Env = append(cmd.Env, "DOSEMU_QUIET=1")
+
+	// Start dosemu with a PTY for terminal-mode I/O. Dosemu translates DOS
+	// screen output (INT 10h) to ANSI via $_term_color and writes to stdout.
+	// The PTY bridges this to the SSH session.
+	ptmx, err := pty.StartWithSize(cmd, &pty.Winsize{Rows: 25, Cols: 80})
+	if err != nil {
+		return fmt.Errorf("failed to start dosemu2 with pty: %w", err)
+	}
+
+	// Set PTY master to raw mode for clean passthrough of CP437 bytes
+	fd := int(ptmx.Fd())
+	if oldState, err := term.MakeRaw(fd); err == nil {
+		defer term.Restore(fd, oldState)
+	}
+
+	// Set up a read interrupt so we can cleanly stop the input goroutine
+	// when the door exits, preventing it from consuming the next keypress.
+	// Sessions that support SetReadInterrupt (SSH) get clean cancellation;
+	// others (telnet) fall back to the old behavior where the input goroutine
+	// exits on its own when ptmx is closed and the next write fails.
+	readInterrupt := make(chan struct{})
+	hasInterrupt := false
+	if ri, ok := ctx.Session.(interface{ SetReadInterrupt(<-chan struct{}) }); ok {
+		ri.SetReadInterrupt(readInterrupt)
+		defer ri.SetReadInterrupt(nil)
+		hasInterrupt = true
+	}
+
+	// Bidirectional I/O: SSH session ↔ PTY master
+	//
+	// Two modes depending on whether a FOSSIL driver is configured:
+	//
+	// FOSSIL mode: $_com1="virtual" maps COM1 to the PTY. The door's output
+	// flows through the FOSSIL driver → COM1 → PTY. $_video="none" suppresses
+	// DOS screen output, but dosemu still writes plain-ASCII boot text directly
+	// to the PTY. Gate on the first ESC (0x1B) byte: boot text is plain ASCII,
+	// while door output (via FOSSIL) contains ANSI escape sequences.
+	//
+	// Terminal mode: -t translates DOS screen writes (INT 10h) to ANSI on
+	// the PTY. The output goroutine buffers until it detects ESC[2J (clear
+	// screen from `cls` in EXTERNAL.BAT), discarding DOS boot text before it.
+	useFossil := ctx.Config.FossilDriver != ""
+
+	inputDone := make(chan struct{})
+	outputDone := make(chan struct{})
+	go func() {
+		defer close(inputDone)
+		_, err := io.Copy(ptmx, ctx.Session)
+		if err != nil && err != io.EOF && !errors.Is(err, os.ErrClosed) {
+			log.Printf("WARN: Node %d: Error copying session stdin to dosemu PTY: %v", ctx.NodeNumber, err)
+		}
+	}()
+	go func() {
+		defer close(outputDone)
+		buf := make([]byte, 4096)
+
+		if useFossil {
+			// FOSSIL mode: dosemu writes plain-ASCII boot text to the PTY before
+			// the door starts. Door output (via FOSSIL → COM1 → PTY) contains
+			// ANSI escape codes. Gate on the first ESC (0x1B) byte to suppress
+			// boot text while forwarding all door output.
+			log.Printf("DEBUG: Node %d: FOSSIL mode — gating boot text until first ESC byte", ctx.NodeNumber)
+			var accum []byte
+			gated := true
+			for {
+				n, err := ptmx.Read(buf)
+				if n > 0 {
+					if gated {
+						accum = append(accum, buf[:n]...)
+						if idx := bytes.IndexByte(accum, 0x1b); idx >= 0 {
+							gated = false
+							ctx.Session.Write(accum[idx:])
+							accum = nil
+							log.Printf("DEBUG: Node %d: FOSSIL boot text gate opened (ESC byte at offset %d)", ctx.NodeNumber, idx)
+						}
+						if gated && len(accum) > 32768 {
+							log.Printf("WARN: Node %d: FOSSIL gate: no ESC after 32KB, flushing", ctx.NodeNumber)
+							gated = false
+							ctx.Session.Write(accum)
+							accum = nil
+						}
+					} else {
+						ctx.Session.Write(buf[:n])
+					}
+				}
+				if err != nil {
+					if gated && len(accum) > 0 {
+						log.Printf("WARN: Node %d: FOSSIL boot text gate never opened. Accumulated %d bytes", ctx.NodeNumber, len(accum))
+					}
+					if err != io.EOF && !errors.Is(err, os.ErrClosed) {
+						log.Printf("WARN: Node %d: Error copying dosemu PTY to session: %v", ctx.NodeNumber, err)
+					}
+					return
+				}
+			}
+		}
+
+		// Terminal mode: gate until ESC[2J (cls) to suppress boot text
+		clearScreen := []byte("\x1b[2J")
+		var accum []byte
+		gated := true
+
+		for {
+			n, err := ptmx.Read(buf)
+			if n > 0 {
+				if gated {
+					accum = append(accum, buf[:n]...)
+					if idx := bytes.Index(accum, clearScreen); idx >= 0 {
+						// Found clear screen — forward it and everything after
+						gated = false
+						ctx.Session.Write(accum[idx:])
+						accum = nil
+						log.Printf("DEBUG: Node %d: DOS boot text gate opened (clear screen detected)", ctx.NodeNumber)
+					}
+					// Safety: if we've accumulated too much without finding cls, flush it
+					if gated && len(accum) > 32768 {
+						log.Printf("WARN: Node %d: Clear screen not found after 32KB, flushing", ctx.NodeNumber)
+						gated = false
+						ctx.Session.Write(accum)
+						accum = nil
+					}
+				} else {
+					ctx.Session.Write(buf[:n])
+				}
+			}
+			if err != nil {
+				if gated && len(accum) > 0 {
+					log.Printf("WARN: Node %d: DOS boot text gate never opened. Accumulated %d bytes. First 200: %q", ctx.NodeNumber, len(accum), accum[:min(200, len(accum))])
+				}
+				if err != io.EOF && !errors.Is(err, os.ErrClosed) {
+					log.Printf("WARN: Node %d: Error copying dosemu PTY to session: %v", ctx.NodeNumber, err)
+				}
+				return
+			}
+		}
+	}()
+
+	// Wait for dosemu to exit, then cleanly shut down I/O goroutines
+	cmdErr := cmd.Wait()
+	log.Printf("DEBUG: Node %d: dosemu2 process exited for door '%s'", ctx.NodeNumber, ctx.DoorName)
+
+	// Interrupt the input goroutine's blocked Read() so it exits without
+	// consuming the user's next keypress, then close the PTY.
+	close(readInterrupt)
+	if hasInterrupt {
+		<-inputDone
+	}
+	ptmx.Close()
+	<-outputDone
+
+	if cmdErr != nil {
+		log.Printf("ERROR: Node %d: DOS door '%s' failed: %v", ctx.NodeNumber, ctx.DoorName, cmdErr)
+	} else {
+		log.Printf("INFO: Node %d: DOS door '%s' completed successfully", ctx.NodeNumber, ctx.DoorName)
+	}
+
+	// Run cleanup while dropfiles/node dirs still exist (before deferred cleanup fires)
+	executeCleanup(ctx)
+
+	return cmdErr
+}
